app/tg_bot/requirements.txt
aiogram==2.24
aiohttp
python-dotenv
pyairtable
